requirements.txt
apt-xapian-index==0.49
autopep8==1.5.4
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==4.3.1
entrypoints==0.3
evdev==1.3.0
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
img2pdf==0.3.3
importlib-resources==3.0.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.1.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pbr==5.5.0
pexpect==4.6.0
Pillow==7.0.0
psutil==5.5.1
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sltx==0.1.5
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
websocket-client==0.57.0
xkit==0.0.0
